// socket/socket.enum.ts
export enum Events {
    ERROR = 'error',
    USERS = 'users',
    MESSAGE = 'message',
    CONNECTION = 'connection',
    DISCONNECTION = 'disconnect',
    ORDER = 'orders',
    LANCER = "lancer",
    NEW_COUNT = "new_count",
    ONLINE_USERS = "online_users"
}

export enum REDIS {
    USER_LIST = "userlist"
}